backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
fastapi==0.135.1
fonttools==4.60.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.9.2
mne==1.8.0
numpy==1.26.4
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.22
requests==2.32.5
scikit-learn==1.7.2
scipy==1.13.1
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.42.0
